pyproject.toml
[project]
name = "mlforge"
version = "1.0.0"
description = "Autonomous ML research framework"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Robert Lupo"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "gitpython>=3.1",
    "jinja2>=3.1",
    "scikit-learn>=1.5",
    "pandas>=2.0",
    "numpy>=2.0",
    "xgboost>=2.0",
    "lightgbm>=4.0",
    "optuna>=4.0",
    "pyarrow>=15.0",
    "rich>=13.0",
    "psutil>=6.0",
]

[project.urls]
Homepage = "https://github.com/robertlupo1997/mlforge"
Repository = "https://github.com/robertlupo1997/mlforge"
Documentation = "https://github.com/robertlupo1997/mlforge/tree/main/docs"

[project.optional-dependencies]
dl = [
    "torch>=2.4",
    "torchvision>=0.19",
    "timm>=1.0",
    "transformers>=4.45",
    "datasets>=3.0",
]
ft = [
    "peft>=0.14",
    "trl>=0.14",
    "bitsandbytes>=0.45",
    "evaluate>=0.4",
    "rouge-score>=0.1",
    "transformers>=4.45",
    "datasets>=3.0",
]

[project.scripts]
mlforge = "mlforge.cli:main"

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mlforge"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (e.g. scaffold + run train.py)",
]
